Feat: header에 프로필 사진 영역 이동
